fixed bug building level was double should have been long

+++ PurrfectEvolution/app/src/main/java/purrfect/evolution/DataContainerForPurfectEvolution.java
@@ -36,31 +36,31 @@ final class DataContainerForPurfectEvolution {
 
     //Variables for buildings for saving purposes
     int mTypeBuilding1 = 0;
-    double mLevelBuilding1 = 0;
+    long mLevelBuilding1 = 0;
 
     int mTypeBuilding2 = 0;
-    double mLevelBuilding2 = 0;
+    long mLevelBuilding2 = 0;
 
     int mTypeBuilding3 = 0;
-    double mLevelBuilding3 = 0;
+    long mLevelBuilding3 = 0;
 
     int mTypeBuilding4 = 0;
-    double mLevelBuilding4 = 0;
+    long mLevelBuilding4 = 0;
 
     int mTypeBuilding5 = 0;
-    double mLevelBuilding5 = 0;
+    long mLevelBuilding5 = 0;
 
     int mTypeBuilding6 = 0;
-    double mLevelBuilding6 = 0;
+    long mLevelBuilding6 = 0;
 
     int mTypeBuilding7 = 0;
-    double mLevelBuilding7 = 0;
+    long mLevelBuilding7 = 0;
 
     int mTypeBuilding8 = 0;
-    double mLevelBuilding8 = 0;
+    long mLevelBuilding8 = 0;
 
     int mTypeBuilding9 = 0;
-    double mLevelBuilding9 = 0;
+    long mLevelBuilding9 = 0;
 
 
     public DataContainerForPurfectEvolution()
